# specs/001-auth-customer-staff/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dahab Auth API
  version: 1.0.0
  description: |
    Authentication surface for the Dahab Gold Marketplace Backend.
    Two principals — customer and dashboard staff — share primitives
    (rate limits, audit log, error envelope) but do not share tokens.
    All rules trace back to `docs/Technical Spec/dahab-spec-part1-auth.md`.

    Two API surfaces, two guards, four token abilities:

    | Surface | Prefix | Guard | Access ability | Refresh ability |
    |---|---|---|---|---|
    | Customer API | `/customer/*` | `auth:customer` | `customer:access` | `customer:refresh` |
    | Dashboard API | `/dashboard/*` | `auth:staff` | `staff:access` | `staff:refresh` |

    A token of the other principal is rejected with 401 `unauthenticated`. A token
    of the right principal but the wrong ability (a refresh token on an access
    endpoint, or the reverse) is rejected with 403 `forbidden`.

    `x-implemented: true` marks operations that exist in the running API today;
    `x-implemented: false` marks planned operations (see `../tasks.md`).
servers:
  - url: /api/v1
    description: v1
components:
  securitySchemes:
    customerBearer:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
      description: Customer API. The customer access token (ability `customer:access`).
    customerRefreshBearer:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
      description: Customer API refresh endpoint only. The customer refresh token (ability `customer:refresh`).
    dashboardBearer:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
      description: Dashboard API. The staff access token (ability `staff:access`).
    dashboardRefreshBearer:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
      description: Dashboard API refresh endpoint only. The staff refresh token (ability `staff:refresh`).
  schemas:
    Envelope:
      type: object
      required: [data]
      properties:
        data:
          nullable: true
        meta:
          type: object
...
    Session:
      type: object
      required: [access_token, access_token_expires_at, refresh_token, refresh_token_expires_at, token_type]
      properties:
        token_type: { type: string, enum: [Bearer] }
        access_token: { type: string, description: "Sanctum PAT, ~15 min TTL, ability `<principal>:access`" }
        access_token_expires_at: { type: string, format: date-time }
        refresh_token: { type: string, description: "Sanctum PAT with ability `<principal>:refresh`, 30 day TTL, rotates on every use" }
        refresh_token_expires_at: { type: string, format: date-time }
        family_id: { type: string, format: uuid }
    OtpChallenge:
      type: object
      required: [otp_required, challenge_id, expires_at, resend_available_at]
      properties:
        otp_required: { type: boolean, enum: [true] }
        challenge_id: { type: string, format: uuid }
        expires_at: { type: string, format: date-time }
        resend_available_at: { type: string, format: date-time }
...
    Customer:
      type: object
      required: [id, display_ref, phone, is_verified, is_suspended, trade_allowed, preferred_lang, created_at]
      properties:
        id: { type: string, format: uuid }
        display_ref: { type: string }
        phone: { type: string }
        email: { type: string, nullable: true, format: email }
        email_verified_at: { type: string, nullable: true, format: date-time }
        full_name: { type: string, nullable: true }
        preferred_lang: { type: string, enum: [ar, en] }
        is_verified: { type: boolean }
        is_suspended: { type: boolean }
        suspended_reason: { type: string, nullable: true }
        suspended_at: { type: string, nullable: true, format: date-time }
        trade_allowed: { type: boolean }
        created_at: { type: string, format: date-time }
...
  parameters:
    DeviceIdHeader:
      in: header
      name: X-Device-Id
      required: false
      description: Client-generated opaque device id (128-bit random, persisted client-side)
      schema: { type: string, minLength: 16, maxLength: 128 }
    DevicePlatformHeader:
      in: header
      name: X-Device-Platform
      required: false
      schema: { type: string, enum: [ios, android, web] }
paths:
...
  /customer/auth/login:
    post:
      tags: [Customer Auth]
      x-implemented: true
      security: []
      summary: Customer sign-in (phone + password; OTP on new device)
      description: |
        Implemented today: the known-device branch only. A correct password from a device that is
        not trusted is refused with 401 `invalid_credentials` until the OTP branch (US3, tasks
        T081–T091) is built; the `OtpChallenge` response below is the planned shape.
        Identity bucket exhausted → 429 `account_locked`; per-IP bucket exhausted → 429 `too_many_requests`.
      parameters:
        - $ref: '#/components/parameters/DeviceIdHeader'
        - $ref: '#/components/parameters/DevicePlatformHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password]
              properties:
                phone: { type: string }
                password: { type: string, writeOnly: true }
      responses:
        '200':
          description: Session issued (known device) OR OTP challenge (new device)
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: '#/components/schemas/Envelope'
                      - type: object
                        properties:
                          data:
                            type: object
                            properties:
                              customer: { $ref: '#/components/schemas/Customer' }
                              session: { $ref: '#/components/schemas/Session' }
                  - allOf:
                      - $ref: '#/components/schemas/Envelope'
                      - type: object
                        properties:
                          data: { $ref: '#/components/schemas/OtpChallenge' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /customer/auth/refresh:
    post:
      tags: [Customer Auth]
      x-implemented: true
      security:
        - customerRefreshBearer: []
      summary: Rotate the customer refresh token
      description: The bearer token here MUST be a customer refresh token (ability `customer:refresh`); an access token is 403. Replaying an already-rotated refresh token revokes the whole family.
      responses:
        '200':
          description: New access/refresh pair, same family
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/RefreshInvalid' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /customer/auth/logout:
    post:
      tags: [Customer Auth]
      x-implemented: true
      security:
        - customerBearer: []
      summary: Revoke the current access+refresh pair
      responses:
        '204': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /customer/auth/logout-all:
    post:
      tags: [Customer Auth]
      x-implemented: true
      security:
        - customerBearer: []
      summary: Revoke every token for the current actor
      responses:
        '204': { description: Logged out from all devices }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /customer/auth/me:
    get:
      tags: [Customer Auth]
      x-implemented: true
      security:
        - customerBearer: []
      summary: Get the authenticated customer
      responses:
        '200':
          description: Own customer row
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Customer' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...